Use default timeout value from OkHttpClient
For some reason app was using a huge value for timeout (2 minutes)
since of the beginning which causes app not giving any feedback to
the user in a reasonable amount of time when server is really down.

Changed timeout value to the default one, which is 10 seconds
according to current implementation of OkHttpClient.