Convert Firebase timestamp to Long on remote layer
Server side is returning timestamp exactly as Firebase returns
which is an object with seconds and nanoseconds inside. Since
that isn't the common format, some conversions were being carried
to other layers than the remote.

Created an intermediate type (SessionRemoteFirebase) to get results
from server and than convert it to SessionRemote which now uses
timestamp as a Long.