Fix the screen orientation for portrait
There are activities that are not set as portrait and
the app layout does not work in landscape and the app
screens must be fixed in portrait.

For that, add screen orientation for all the app activities
and add user-feature "android.hardware.screen.portrait" for
info that the app work in portrait only.

fixed: #176