Create meetings presenter
Create meetings presenter which gets attached to view and during on
resume calls list meetings use case. Then meetings list is sorted by
date and then converted to a list of meeting items, which is a model
with formatted data ready to be used by View.

Also changed return type of meetings use case and repository to use
Kotlin MutableList, which is compatible with Java, instead of Kotlin
List.