Creates Meetings Activity
Adds meetings Activity and RecycleView adapter. Managed to use only
one layout for meeting card view by defining custom layout attributes
on attrs.xml and themes for both "recent" and "old" types.

Activity is still disconnected from the rest of app's flow and the UI
is uncompleted when compared with planned design, missing elements
include: the menu on left side of toolbard and notifications icon on
the right side.